+++ PWM_GPIO.c
@@ -10,7 +10,7 @@
 #include "TIM_ex.h"
 #include "PWM_GPIO.h"
 
-#if defined(HAL_TIM_MODULE_ENABLED)
+#if defined(HAL_GPIO_MODULE_ENABLED) && defined(HAL_TIM_MODULE_ENABLED)
 /****************************************************************/
 
 
@@ -98,6 +98,6 @@ void NONNULL__ PWM_GPIO_handler(PWM_GPIO * const pPWM)
 }
 
 
-/****************************************************************/
-#endif	/* defined(HAL_TIM_MODULE_ENABLED) */
-/****************************************************************/
+/*******************************************************************************/
+#endif	/* defined(HAL_TIM_MODULE_ENABLED) && defined(HAL_GPIO_MODULE_ENABLED) */
+/*******************************************************************************/